Revise command structure/grouping
